pro-console: fix runtime error when expanding pending/processed scheduled events
closes: https://github.com/hasura/graphql-engine/issues/7168

### Description

On expanding the pending and processed event on `one-off scheduled event` on pro we are getting an error because we are not passing the auth header to API request. In this PR we are passing `includeAdminHeaders` to API as `true` to fetch events.
`keyframes` imported from `styled-components` is supported on V3 only and we are using V5.0.1. So, instead of importing spinner from UIKit which we are not using anymore, we are importing spinner form other place.

### Affected components

- [x] Console

https://github.com/hasura/graphql-engine-mono/pull/1754

Co-authored-by: Sooraj <8408875+soorajshankar@users.noreply.github.com>
GitOrigin-RevId: 7550d544b2d7f4f98a793895368c964d17d0e362