console: make specifying unique params for rate limits optional
### Description
Make the unique parameters to be an optional field in the form.

### Changelog
- [x] `CHANGELOG.md` is updated with user-facing content relevant to this PR. If no changelog is required, then add the `no-changelog-required` label.

### Affected components
- [x] Console

### Related Issues
Based on internal discussion: https://hasurahq.slack.com/archives/C01SQFJ7SHX/p1626948719012400

Closes #1890

### Solution and Design
This is more of adhoc solution for the unique parameters not being a required field: Using a checkbox to enable the additional unique params - if the user has not checked it, then the payload is sent without the unique params attribute.

When the additional unique parameters are not selected -

![Screenshot 2021-07-26 at 2 40 57 PM](https://user-images.githubusercontent.com/11921040/126964312-bb12b146-2644-432f-a790-88f23fb322ce.png)

When the additional unique parameters are picked -

![Screenshot 2021-07-26 at 2 42 03 PM](https://user-images.githubusercontent.com/11921040/126964545-89f41e43-c855-4290-8b15-75f03c44101e.png)

https://github.com/hasura/graphql-engine-mono/pull/1885

GitOrigin-RevId: dd16812c1c548147a47830a8a8af27c1e39a5689