Add Execute part for the MySQL data loader
The Plan part is missing, because it needs support from FromIr. That'll come in a follow up commit.

**Next PR**: #2529

This is the result of splitting up the mega PR into more digestible chunks. This is the smallest subset I've been able to collect. Missing parts are noted in comments.

The code isn't reachable from Main, so it won't affect the test suite. It just gets compiled for now.

For context, this splits up work from https://github.com/hasura/graphql-engine-mono/pull/2332

PR-URL: https://github.com/hasura/graphql-engine-mono/pull/2511
Co-authored-by: Abby Sassel <3883855+sassela@users.noreply.github.com>
GitOrigin-RevId: 00f30b0f494b56b3b7f8c1b0996377db4874c88d