Add backend modules for MySQL
This is the next part in the series of MySQL PRs.

**Purpose**: Adds the modules related to the generic backend abstraction. This is the penultimate PR.

The final PR will be the more substantial change made to FromIr (and small tweaks to Execute) that will connect all the things together including Python tests.

**Preceding PR:** #2529

**Next PR**: #2608

After #2529 is merged, this can be repointed to `main`. For now it's aimed at #2529, because then the diff display is simpler.

PR-URL: https://github.com/hasura/graphql-engine-mono/pull/2549
Co-authored-by: Abby Sassel <3883855+sassela@users.noreply.github.com>
GitOrigin-RevId: 9d48ac38f4ac7c32c79ba521d3de40805c2a13bc