ci: combine pro/oss build jobs
The motivation here is:
- save some CI resources
- make it cleaner/more sensible to do library caching

Since OSS is a dependency of pro, and at the same time the various
executables can be built in parallel, doing this all in the same job
probably is more sensible.

Future improvements: there are separate pro and oss pipelines and one or
the other of those will need to wait unnecessarily while everything
finishes. In practice, currently this isn't a big deal, but we could use
the BuildKite blocking job API to get more fine-grained downstream
triggering.

GitOrigin-RevId: aa917996cbfc678becaaafca490d6edc7de89b1f